requirements.txt
pandas==2.2.3
click==8.2.1
google-genai==1.18.0
pytest==8.3.5
python-dotenv==1.1.0
aiosqlite==0.21.0
pydantic==2.11.5
asyncclick==8.1.8.0
pytest-asyncio==1.0.0